Bug fixes for arrays and subtitles (#136)
* Fixed issues with arrays and subtitles

* Fix subtitle issues with speed and slowing down

* Lint

* Fixed tests

* Removed default animation time

+++ src/main/kotlin/com/manimdsl/linearrepresentation/MObjects.kt
@@ -121,7 +121,7 @@ data class SubtitleBlock(
     val variableNameGenerator: VariableNameGenerator,
     private var boundary: List<Pair<Double, Double>> = emptyList(),
     val textColor: String? = null,
-    var duration: Int,
+    var duration: Double,
     override val runtime: Double = 1.0,
     override val ident: String = variableNameGenerator.generateNameFromPrefix("subtitle_block")
 ) : ShapeWithBoundary("_subtitle") {
